Remove DeleteEpiredProfile from ProfileManager interface as it is App Store Connect API specific, Spaceship returns expired profiles correcty
